Replace "NSInteger" with "NSId" (sqlite3_int64) for all Sqlite3 related indexes. Replace NSInteger with (long)NSInteger for all format strings

+++ Geocube/AppDelegate.m
@@ -269,7 +269,7 @@
 
 - (void)switchController:(NSInteger)idx
 {
-    NSLog(@"AppDelegate: Switching to TB %ld", idx);
+    NSLog(@"AppDelegate: Switching to TB %ld", (long)idx);
     self.window.rootViewController = [tabBars objectAtIndex:idx];
     [self.window makeKeyAndVisible];
 }
