+++ Geocube/CacheTableViewCell.m
@@ -163,7 +163,7 @@
     ratingT.image = [imageLibrary getRating:t];
 
     if (favs != 0) {
-        favourites.text = [NSString stringWithFormat:@"%ld", favs];
+        favourites.text = [NSString stringWithFormat:@"%ld", (long)favs];
         imgFavouritesIV.hidden = FALSE;
     }
 }
